printserver_win/requirements.txt
flask>=3.0.0
werkzeug>=3.0.0
waitress>=3.0.0
pywin32>=306
qrcode>=7.4.2
pyserial>=3.5
